--- src/main/java/com/menu/pubganalyzer/controller/GlobalExceptionHandler.java
@@ -1,77 +1,0 @@
-package com.menu.pubganalyzer.controller;
-
-import com.menu.pubganalyzer.exception.MatchNotFoundException;
-import com.menu.pubganalyzer.exception.ParticipantNotFoundException;
-import com.menu.pubganalyzer.exception.PlayerNotFoundException;
-import com.menu.pubganalyzer.support.apiResult.ApiResult;
-import com.menu.pubganalyzer.support.apiResult.ApiResultUtil;
-import com.menu.pubganalyzer.util.pubgAPI.exception.PubgAPIPlayerNotFoundException;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.ui.Model;
-import org.springframework.validation.BindException;
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Slf4j
-@ControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(value = {
-            ParticipantNotFoundException.class,
-            MatchNotFoundException.class,
-            PlayerNotFoundException.class
-    })
-    public String badRequestExceptionHandler(Throwable e, Model model) {
-        model.addAttribute("viewTitle", "잘못된 접근");
-        model.addAttribute("errMsg", e.getMessage());
-
-        return "error/bad-request";
-    }
-
-    @ExceptionHandler(value = BindException.class)
-    public String validationExceptionHandler(BindException e, Model model) {
-        BindingResult bindingResult = e.getBindingResult();
-        Map<String, String> result = new HashMap<>();
-
-        String key;
-
-        for (FieldError fieldError : bindingResult.getFieldErrors()) {
-            key = fieldError.getField();
-            if (key.lastIndexOf('.') != -1) {
-                key = key.substring(key.lastIndexOf('.') + 1);
-            }
-            result.put(key, fieldError.getDefaultMessage());
-        }
-
-        model.addAttribute("viewTitle", "잘못된 접근");
-        model.addAttribute("errMsg", result.toString());
-
-        return "error/bad-request";
-    }
-
-    @ResponseBody
-    @ExceptionHandler(value = {
-            PubgAPIPlayerNotFoundException.class
-    })
-    public ResponseEntity<ApiResult<?>> playerNotFoundExceptionHandler(Throwable throwable) {
-        return errorResponse(throwable, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(Exception.class)
-    public String exceptionHandler(Exception e) {
-        log.error("{}", e.getMessage(), e);
-        return "error/error";
-    }
-
-    private ResponseEntity<ApiResult<?>> errorResponse(Throwable throwable, HttpStatus status) {
-        return ResponseEntity.status(status)
-                .body(ApiResultUtil.error(status.value(), throwable.getMessage()));
-    }
-}
